Split tests into a separate file
